pyproject.toml
[tool.poetry]
name = "shibabot"
version = "0.1.0"
description = "Chatbot for Discord"
authors = ["Todd Birchard <toddbirchard@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/toddbirchard/shibabot/"
repository = "https://github.com/toddbirchard/shibabot/"
documentation = "https://github.com/toddbirchard/shibabot/"
keywords = [
    "Discord",
    "Bot",
    "Chatbot",
    "Python",
    "Discord Bot"
]

[tool.poetry.dependencies]
python = "^3.8"
discord = "*"
requests = "*"
loguru = "*"
pandas = "*"
plotly = "*"
chart-studio = "*"
wikipedia-api = "*"
imdbpy = "*"
emoji = "*"
simplejson = "*"
python-dotenv = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
mypy = "*"
black = "^20.8b1"
isort = "^5.6.4"

[tool.poetry.scripts]
run = "wsgi:start"

[tool.poetry.urls]
issues = "https://github.com/toddbirchard/shibabot/issues"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

